rename watch service
