+++ packages/cli/src/directory-process-service/directory-process-service.ts
