+++ packages/cli/src/index.ts
@@ -1,2 +1,3 @@
 export * from './build';
 export * from './base-generator';
+export * from './directory-process-service/directory-process-service';
